Refactor directory structure and improve code organization
Refactor the codebase to improve structure and address potential bugs.

* Add `ConfigManager` class in `config_manager.py` to manage application configuration.
* Add `EncryptionManager` class in `encryption_manager.py` to handle encryption.
* Add `DatabaseManager` class in `database_manager.py` to handle database operations.
* Move `generate_key.py` to `utils` directory.
* Move `config.json` to `config` directory.
* Move `face_recognition_model.pkl`, `classifier.xml`, and `haarcascade_frontalface_default.xml` to `models` directory.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Ayaan2907/Visual-Attendence-Tracking?shareId=XXXX-XXXX-XXXX-XXXX).
